% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{amdMataxpy}
\alias{amdMataxpy}
\title{AMD GPU Matrix Multiplication}
\usage{
amdMataxpy(alpha, A, B)
}
\arguments{
\item{alpha}{Numeric value to multiply the A matrix}

\item{A}{A amdMatrix object}

\item{B}{A amdMatrix object}
}
\value{
A amdMatrix or gpuMatrix (if data type is 'integer') object
}
\description{
matrix multiplication
}
\author{
Charles Determan Jr.
}

