% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{amdMatMult}
\alias{amdMatMult}
\title{AMD GPU Matrix Multiplication}
\usage{
amdMatMult(A, B)
}
\arguments{
\item{A}{A amdMatrix object}

\item{B}{A amdMatrix object}
}
\value{
A amdMatrix object
}
\description{
matrix multiplication
}
\note{
Please note that for integer matrices there are no defined BLAS
routines.  As such, any 'integer' type matrices will currently deprecate
to the \link[gpuR]{gpuMatrix} class.
}
\author{
Charles Determan Jr.
}

