% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{amdMatrix-class}
\alias{amdMatrix-class}
\title{amdMatrix Class}
\description{
This is the 'mother' class for all
amdMatrix objects.  It is a child class from the
gpuMatrix class from gpuR.  All other 
amdMatrix classes inherit from this class but 
there are no current circumstances where this class 
is used directly.

There are multiple child classes that correspond
to the particular data type contained.  These include
\code{iamdMatrix}, \code{famdMatrix}, and 
\code{damdMatrix} corresponding to integer, float, and
double data types respectively.
}
\note{
R does not contain a native float type.  As such,
the matrix data within a \code{\link{famdMatrix-class}} 
will be represented as double but downcast when any 
amdMatrix methods are used.

May also remove the type slot
}
\section{Slots}{

 Common to all amdMatrix objects in the package
 \describe{
     \item{\code{x}:}{An R matrix object}
     \item{\code{type}:}{Character object specifying
     the type the matrix data will be interpreted as}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{iamdMatrix-class}}, 
\code{\link{famdMatrix-class}},
\code{\link{damdMatrix-class}}
}

