% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{famdMatrix-class}
\alias{famdMatrix-class}
\title{famdMatrix Class}
\description{
An integer type matrix in the S4 \code{amdMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{x}:}{A numeric R matrix.}
     \item{\code{type}:}{Character object specifying
     the type the matrix data is intepreted as float}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{amdMatrix-class}},
\code{\link{iamdMatrix-class}},
\code{\link{damdMatrix-class}}
}

