% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{deepcopy,amdMatrix-method}
\alias{deepcopy,amdMatrix-method}
\title{Copy a "gpuRclblas" object}
\usage{
\S4method{deepcopy}{amdMatrix}(object)
}
\arguments{
\item{object}{A gpuRclblas object}
}
\value{
A gpuRclblas object
}
\description{
This is needed to make a duplicate of a gpuRclblas object
}
\details{
This is needed to make a duplicate of a gpuRclblas object 
(i.e. \code{\link{amdMatrix}}, \code{\link{amdVector}}, 
\code{\link{vclAMDMatrix}}, \code{\link{vclAMDVector}} because
the traditional syntax would only copy the pointer of the object.
}
\author{
Charles Determan Jr.
}

