\name{gpuRclblas-package}
\alias{gpuRclblas-package}
\alias{gpuRclblas}
\docType{package}
\title{
\packageTitle{gpuRclblas}
}
\description{
\packageDescription{gpuRclblas}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gpuRclblas}
\packageIndices{gpuRclblas}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gpuRclblas}

Maintainer: \packageMaintainer{gpuRclblas}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
