% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{iamdMatrix-class}
\alias{iamdMatrix-class}
\title{iamdMatrix Class}
\description{
An integer type matrix in the S4 \code{amdMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{x}:}{A integer typed R matrix}
     \item{\code{type}:}{Character object specifying
     the type the matrix data is integer}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{amdMatrix-class}}, 
\code{\link{iamdMatrix-class}},
\code{\link{damdMatrix-class}}
}

